systemd $config_file param (PEAUTO-330)
Adds a `$config_file` parameter, used on `systemd` distros to optionally specify
which config file to load.

In our environment, we deploy two separate Vault configs: a "current" one,
which is the live config Vault is currently running on, and an
"up-to-date" one, sitting there waiting for a human to decide when it
should be promoted to current.

On Debian/Ubuntu, we use `/etc/default/vault` to specify which config
file Vault should use.

On `systemd` distros an equivalent mechanism didn't exist, resulting in
all `*.json` and `*.hcl` files under `$config_dir` to be loaded.